+++ NEWS.md
@@ -2,6 +2,7 @@
 * released: ???
 * implemented str.fsets(), as_tibble.fsets()
 * removed dependency on the zoo package
+* added lowerBound() algebra for missing values
 
 # lfl 2.2.1
 * released: 2024-10-22